pyproject.toml
[project]
name = "global-air-quality-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime>=5.5",
    "folium>=0.20.0",
    "marimo>=0.14.10",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "streamlit>=1.46.1",
    "streamlit-folium>=0.25.0",
]
[tool.pytest.ini_options]
pythonpath = ["src"]